Stats y EndGame
Se han reincorporado las stats y se ha añadido transición de pantalla al acabar la partida al igual que en el resto de pantallas.

+++ HitOnTheBeat/Assets/Scripts/PlayerSelector.cs
@@ -10,5 +10,5 @@ public class PlayerSelector : MonoBehaviour
     public int jumpStats;
     public int pushStats;
     public int killsStats;
-    public float averageRhythmStats;
+    public int averageRhythmStats;
 }
